refactor: seperate handler and business logic into different files
