--- backend/internal/auth/login.go
@@ -1,69 +1,0 @@
-package auth
-
-import (
-	"encoding/json"
-	"net/http"
-	"os"
-	"time"
-
-	"github.com/golang-jwt/jwt/v5"
-)
-
-type LoginRequest struct {
-	Username string `json:"user"`
-	Password string `json:"password"`
-}
-
-type LoginResponse struct {
-	Token string
-}
-
-var jwtKey = []byte(os.Getenv("BACKEND_SECRET_KEY"))
-
-type Claims struct {
-	Username string `json:"username"`
-	jwt.RegisteredClaims
-}
-
-func LoginHandler(w http.ResponseWriter, r *http.Request) {
-	var u LoginRequest
-
-	// Decode the request body into the LoginRequest struct
-	err := json.NewDecoder(r.Body).Decode(&u)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-	defer r.Body.Close()
-
-	// Call the Login function
-	err = Login(u)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusUnauthorized)
-		return
-	}
-
-	// Login successful, generate a token
-	expirationTime := time.Now().Add(120 * time.Hour)
-	claims := &Claims{
-		Username: u.Username,
-		RegisteredClaims: jwt.RegisteredClaims{
-			ExpiresAt: jwt.NewNumericDate(expirationTime),
-		},
-	}
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	tokenString, err := token.SignedString(jwtKey)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	// Return the token as a JSON response
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(LoginResponse{Token: tokenString})
-}
-
-func Login(u LoginRequest) error {
-	// TODO: implement login with database
-	return nil
-}
