+++ backend/internal/auth/service.go
@@ -1,0 +1,56 @@
+package auth
+
+import (
+	"errors"
+	"os"
+	"time"
+
+	"github.com/golang-jwt/jwt/v5"
+)
+
+var jwtKey = []byte(os.Getenv("BACKEND_SECRET_KEY"))
+
+type Claims struct {
+	Username string `json:"username"`
+	jwt.RegisteredClaims
+}
+
+func Register(u RegisterRequest) (string, error) {
+	// TODO: Implement registration with the database
+
+	// Registration successful, generate a token
+	expirationTime := time.Now().Add(120 * time.Hour)
+	claims := &Claims{
+		Username: u.Username,
+		RegisteredClaims: jwt.RegisteredClaims{
+			ExpiresAt: jwt.NewNumericDate(expirationTime),
+		},
+	}
+	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
+	tokenString, err := token.SignedString(jwtKey)
+	if err != nil {
+		return "", errors.New("error_generating_token")
+	}
+
+	return tokenString, err
+}
+
+func Login(u LoginRequest) (string, error) {
+	// TODO: implement login with database
+
+	// Registration successful, generate a token
+	expirationTime := time.Now().Add(120 * time.Hour)
+	claims := &Claims{
+		Username: u.Username,
+		RegisteredClaims: jwt.RegisteredClaims{
+			ExpiresAt: jwt.NewNumericDate(expirationTime),
+		},
+	}
+	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
+	tokenString, err := token.SignedString(jwtKey)
+	if err != nil {
+		return "", errors.New("error_generating_token")
+	}
+
+	return tokenString, err
+}
